Rewrite tests/helperServer: TS and promises
