+++ apps/prairielearn/src/tests/permissions/courseAdminAccess.test.js
@@ -1,4 +1,3 @@
-const util = require('util');
 const assert = require('chai').assert;
 const { step } = require('mocha-steps');
 const { config } = require('../../lib/config');
@@ -79,9 +78,7 @@ function runTest(context) {
 
   var new_user = 'garbage@illinois.edu';
 
-  before('set up testing server', async function () {
-    await util.promisify(helperServer.before().bind(this))();
-  });
+  before('set up testing server', helperServer.before().bind(this));
 
   before('insert users and make instructor course owner', async function () {
     for (const user of users) {
