+++ apps/prairielearn/src/tests/showClosedAssessmentScore.test.js
@@ -1,4 +1,3 @@
-const util = require('util');
 const assert = require('chai').assert;
 const { step } = require('mocha-steps');
 
@@ -28,7 +27,7 @@ describe('Exam assessment with showClosedAssessment AND showClosedAssessmentScor
   };
 
   before('set up testing server', async function () {
-    await util.promisify(helperServer.before().bind(this))();
+    await helperServer.before().call(this);
     const results = await sqldb.queryOneRowAsync(sql.select_exam9, []);
     context.assessmentId = results.rows[0].id;
     context.assessmentUrl = `${context.courseInstanceBaseUrl}/assessment/${context.assessmentId}/`;
